Make the Name fileds in Common Database unique
We want to make sure that the names of the categorys and symbols
of the Common Database are unique. THis is done in SQLite with
Indexes.
The following commands were done in the database since the indexes
were not created at creation time:
CREATE UNIQUE INDEX symbol_name ON Symbol(Name)
CREATE UNIQUE INDEX category_name ON Category(Name)