Repository level added
Created the concept of *Identity classes. These contain the
information needed to find an item in either the local or common
databases.

Used Parcel to be able to pass the *Identity from fragment to
fragment.

While this seems to work there is something I don't understand. I
can use viewModel.viewScope.launch in Category and Meaning but in
SymbolFragment I couldn't use it. I had to use lifecycle.coroutineScope.launchI am not sure why, but I am
saving this now because it seems to work.

I really don't like using
ViewModel.viewModelScopelaunch(Dispatch.IO) in the Fragment layer.
I think it would be better to do that in the viewModel layer. You
will see load*List() routines as an attempt to try this. It seemed
to work, but then I started getting errors aboout can't touch a
view from a thread that didn't create the view. I went back to
passing the flow back to the fragment layer and it too started
havingthe same error about threds. Finally I got it working again
with ViewModel.viewScope.launch in Category and Meaning fragment
but Symbol fragment needed lifecycle.coroutineScope.launch().

Very confusing