Implemented AddSymbol to local database
I can click on Create New Symbol in drop down and add a symbol.
It shows up in the All listing and in the category chosen. If a new
category was entered, the new category shows up and it has the new
symbol listed. Clicking on the symbol in the category list will
show the category.